Updating with testing parity for old/new parser versions
